+++ bot/docker-entrypoint.sh
@@ -1,12 +1,12 @@
 #!/bin/sh
 set -e
 
-if [[ -z $TELEGRAM_TOKEN ]]; then
+if [[ -z "${TELEGRAM_TOKEN}" ]]; then
 	echo "A TELEGRAM_TOKEN is required to run this container."
 	exit 1
 fi
 
-if [[ -z $URI_MONGODB ]]; then
+if [[ -z "${URI_MONGODB}" ]]; then
 	echo "A URI_MONGODB is required to run this container."
 	exit 1
 fi
@@ -14,3 +14,4 @@ fi
 /usr/sbin/crond -b -l 9
 
 exec "$@"
+